pyproject.toml
[project]
name = "temp"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "pikamonvvs", email = "pikamonvvs@naver.com"},
]
dependencies = ["watchdog>=6.0.0", "logger>=1.4", "loguru>=0.7.3", "anyio>=4.9.0", "ffmpeg-python>=0.2.0", "httpx[http2]>=0.28.1", "jsengine>=1.0.7.post1", "streamlink>=7.1.3", "httpx-socks>=0.10.0", "jsonpath-ng>=1.7.0", "types-PyYAML>=6.0.12.20250326", "PyYAML>=6.0.2"]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
